+++ src/core/config/Config_default.h
@@ -57,6 +57,7 @@ R"===(
         "huge-pages": true,
         "hw-aes": null,
         "priority": null,
+        "max-threads-hint": 100,
         "asm": true,
         "argon2-impl": null,
         "cn/0": false,
@@ -76,6 +77,7 @@ R"===(
     "pools": [
         {
             "algo": null,
+            "coin": null,
             "url": "donate.v2.xmrig.com:3333",
             "user": "YOUR_WALLET_ADDRESS",
             "pass": "x",