+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v4.2.0-beta
+- [#1202](https://github.com/xmrig/xmrig/issues/1202) Fixed algorithm verification in donate strategy.
+- Added per pool option `coin` with single possible value `monero` for pools without algorithm negotiation, for upcoming Monero fork.
+- Added config option `cpu/max-threads-hint` and command line option `--cpu-max-threads-hint`.
+
 # v4.1.0-beta
 - **OpenCL backend disabled by default.**.
 - [#1183](https://github.com/xmrig/xmrig/issues/1183) Fixed compatibility with systemd.